+++ src/features/photo/hooks/usePostAnalysisUploadUrl.ts
@@ -1,5 +1,5 @@
 import { useMutation } from "@tanstack/react-query"
-import { postAnalysisUploadUrl } from "../api/presigned-url.api"
+import { postAnalysisUploadUrl } from "../api/image-analysis.api"
 
 export const usePostAnalysisUploadUrl = () => {
   return useMutation({
